Optionally return list of device nodes in Allocate() call
It is not strictly necessary to return the list of device nodes in order
to trigger the NVIDIA container stack to injet a set of GPUs into a
container. However, without this list, the container runtime (i.e.
docker) will not be aware of the device nodes injected into the
container by the NVIDIA container stack.

This normally has little to no consequence for containers relying on the
NVIDIA contaienr stack to allocatn GPUs. However, problems arise
when using GPUs in conjunction with the kubernetes CPUManager. The
following issue summarizes the problem well:

https://github.com/NVIDIA/nvidia-docker/issues/966

With this patch, we add a flag to optionally pass back the list of
device nodes that the NVIDIA container stack will inject into the
container so that the kubelet can forward it to the container runtime.
With this small change, the above issue no longer gets triggered.

Signed-off-by: Kevin Klues <kklues@nvidia.com>